Catch EAGAIN while sending
Seriously, this should not be happening. If we get the "w" flag it means we are
allowed to write. If this then fails, I don't know what's wrong. But we can
easily catch it and then just try again later. It will probably work at some
points (let's hope it does!!)